+++ com.unity.cinemachine/Runtime/Components/CinemachineHardLockToTarget.cs
@@ -45,7 +45,7 @@ namespace Unity.Cinemachine
                 return;
 
             Vector3 dampedPos = FollowTargetPosition;
-            if (deltaTime >= 0)
+            if (VirtualCamera.PreviousStateIsValid && deltaTime >= 0)
                 dampedPos = m_PreviousTargetPosition + VirtualCamera.DetachedFollowTargetDamp(
                     dampedPos - m_PreviousTargetPosition, Damping, deltaTime);
             m_PreviousTargetPosition = dampedPos;