+++ com.unity.cinemachine/Runtime/Behaviours/CinemachineBrain.cs
@@ -615,8 +615,11 @@ namespace Unity.Cinemachine
             CameraUpdateManager.s_CurrentUpdateFilter = updateFilter;
         }
         
-        /// Get the highest-priority Enabled ICinemachineCamera that is in my Channel Mask.
-        ICinemachineCamera TopCameraFromPriorityQueue()
+        /// <summary>
+        /// Choose the default active camera in the case that there is no camera override.
+        /// </summary>
+        /// <returns>The highest-priority Enabled ICinemachineCamera that is in my Channel Mask.</returns>
+        protected virtual ICinemachineCamera TopCameraFromPriorityQueue()
         {
             int numCameras = CameraUpdateManager.VirtualCameraCount;
             for (int i = 0; i < numCameras; ++i)